+++ src/components/Trade.tsx
@@ -8,9 +8,7 @@ import Typography from '@mui/material/Typography';
 import {TradeMemo, TradeState} from "../../apps-script/TradeMemo";
 import {Config} from "../../apps-script/Store";
 import {
-  ChartOptions,
-  createChart,
-  DeepPartial,
+  ChartOptions, createChart, DeepPartial,
   IChartApi,
   ISeriesApi,
   LineStyle
@@ -30,6 +28,7 @@ export default function Trade(props) {
   const [profitLine, setProfitLine] = useState<ISeriesApi<"Line">>(null);
   const [limitLine, setLimitLine] = useState<ISeriesApi<"Line">>(null);
   const [orderLine, setOrderLine] = useState<ISeriesApi<"Line">>(null);
+  const [soldPriceLine, setSoldPriceLine] = useState<ISeriesApi<"Line">>(null);
 
   const map = (prices: number[], mapFn: (v: number) => number) => {
     return prices.map((v, i) => ({time: `${2000 + i}-01-01`, value: mapFn(v)}));
@@ -56,6 +55,7 @@ export default function Trade(props) {
       setLimitLine(chart.current.addLineSeries({color: "red", lineWidth: 1}));
       setProfitLine(chart.current.addLineSeries({color: profitLineColor, lineWidth: 1}))
       setOrderLine(chart.current.addLineSeries({color: "gold", lineWidth: 1}))
+      setSoldPriceLine(chart.current.addLineSeries({color: "cyan", lineWidth: 1}))
     }
 
     chart.current.timeScale().setVisibleLogicalRange({from: 0.5, to: tradeMemo.prices.length - 1.5});
@@ -65,7 +65,7 @@ export default function Trade(props) {
       chart.current = null;
     };
 
-  }, [tradeMemo.prices.length, tradeMemo.getState()]);
+  }, [tradeMemo.prices.length]);
 
   // refresh chart
   useEffect(() => {
@@ -95,12 +95,17 @@ export default function Trade(props) {
 
     if (profitLine) {
       profitLine.applyOptions({
-        visible: !!orderPrice,
+        visible: !!orderPrice && !tradeMemo.stateIs(TradeState.SOLD),
         color: profitLineColor,
         // make dashed if config SellAtTakeProfit is false or HODLing
         lineStyle: !config.SellAtTakeProfit || tradeMemo.hodl ? LineStyle.Dashed : LineStyle.Solid
       });
       profitLine.setData(map(tradeMemo.prices, () => orderPrice * (1 + config.TakeProfit)))
+    }
+
+    if (soldPriceLine) {
+      soldPriceLine.applyOptions({visible: tradeMemo.stateIs(TradeState.SOLD)});
+      soldPriceLine.setData(map(tradeMemo.prices, () => tradeMemo.tradeResult.price))
     }
 
   }, [theme, tradeMemo, config, priceLine, profitLine, limitLine, orderLine]);
@@ -180,29 +185,30 @@ export default function Trade(props) {
           <CardContent>
             <Typography gutterBottom variant="h5" component="div">{props.name}</Typography>
             <Box width={chartOpts.width} height={chartOpts.height} ref={chartContainerRef} className="chart-container"/>
-            <Typography variant="body2" color="text.secondary">
-              <div>Total: {tradeMemo.tradeResult.paid.toFixed(2)}</div>
-              <div>{curProfit > 0 ? "Profit" : "Loss"}: {curProfit.toFixed(2)} ({curProfitPercent}%)</div>
-              <div>Stop: {stopLimitLoss.toFixed(2)} ({stopLimitLossPercent}%)</div>
-            </Typography>
           </CardContent>
+          <Typography marginLeft={"16px"} variant="body2" color="text.secondary">
+            <div>Qty: {tradeMemo.tradeResult.quantity} Paid: {tradeMemo.tradeResult.paid.toFixed(2)}</div>
+            <div>{curProfit > 0 ? "Profit" : "Loss"}: {curProfit.toFixed(2)} ({curProfitPercent}%)</div>
+            <div>Stop: {stopLimitLoss.toFixed(2)} ({stopLimitLossPercent}%)</div>
+          </Typography>
           <CardActions>
             <Stack direction={"row"} spacing={1}>
               {tradeMemo.stateIs(TradeState.BOUGHT) &&
-                <>
-                  <Button size="small" disabled={isSelling} onClick={onSell}>{isSelling ? '...' : 'Sell'}</Button>
-                  <Button size="small" disabled={isBuying} onClick={onBuyMore}>{isBuying ? '...' : 'Buy More'}</Button>
-                  <Box sx={{position: 'relative'}}>
-                    <ToggleButton size="small" value="check" selected={isHodl} color="primary" onChange={flipHodl}
-                                  disabled={isHodlSwitching}>HODL</ToggleButton>
-                    {isHodlSwitching && circularProgress}
-                  </Box>
-                </>
+                <Button size="small" disabled={isSelling} onClick={onSell}>{isSelling ? '...' : 'Sell'}</Button>
               }
-
-              {tradeMemo.stateIs(TradeState.SOLD) || tradeMemo.stateIs(TradeState.BUY) ?
+              {[TradeState.BOUGHT, TradeState.SOLD].includes(tradeMemo.getState()) &&
+                <Button size="small" disabled={isBuying} onClick={onBuyMore}>
+                  {isBuying ? '...' : `Buy ${tradeMemo.stateIs(TradeState.BOUGHT) ? 'More' : 'Again'}`}</Button>
+              }
+              {tradeMemo.stateIs(TradeState.BOUGHT) &&
+                <Box sx={{position: 'relative'}}>
+                  <ToggleButton size="small" value="check" selected={isHodl} color="primary" onChange={flipHodl}
+                                disabled={isHodlSwitching}>HODL</ToggleButton>
+                  {isHodlSwitching && circularProgress}
+                </Box>
+              }
+              {tradeMemo.stateIs(TradeState.SOLD) &&
                 <Button size="small" disabled={isRemoving} onClick={onRemove}>{isRemoving ? '...' : 'Remove'}</Button>
-                : <></>
               }
             </Stack>
           </CardActions>