+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+# v2.1.0
+
+* Trading tab removed, the "Coin Name" input and the "Buy" button moved to the "Assets" tab.
+* Added badges with counters to the state toggle group on "Assets" page. Trade card:
+* Added price at which the asset was sold.
+* Added "Buy Again" button.
+* Added displaying of quantity (Qty).
+* Fixed #7 "Weird javascript rounding issue in Settings tab for Profit / Loss limits"
+* Fixed #9 "Trade result quantity is sometimes rounded to wrong number resulting in error"
+
 # v2.0.1
 
 * Fixed #5 "Regression in profit and commission calc"